将sln改名为TG.UBP.ABP+Zero，表示该sln是基于Abp和Zero的。

+++ src/02-领域层/TG.UBP.Domain.Entity/BaseManage/Permission/ModuleOperate.cs
@@ -14,7 +14,7 @@ namespace TG.UBP.Domain.Entity.BaseManage.Permission
         public string OperateName { get; set; }
 
         [Display(Name = "所属模块")]
-        public int ModuleId { get; set; }
+        public long ModuleId { get; set; }
 
         [Display(Name = "是否验证")]
         public bool IsValid { get; set; }
