+++ src/03-应用层/TG.UBP.Application.Dto/BaseManage/Permission/Modules/ModuleOperateDto.cs
@@ -18,7 +18,7 @@ namespace TG.UBP.Application.Dto.BaseManage.Permission.Modules
         public string OperateName { get; set; }
 
         [Display(Name = "所属模块")]
-        public int ModuleId { get; set; }
+        public long ModuleId { get; set; }
 
         [Display(Name = "是否验证")]
         public bool IsValid { get; set; }
